# packages/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "GridConfig": {
        "additionalProperties": false,
        "properties": {
          "cols": {
            "default": 12,
            "title": "Cols",
            "type": "integer"
          },
          "rowHeight": {
            "default": 40,
            "title": "Rowheight",
            "type": "integer"
          }
        },
        "title": "GridConfig",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LayoutPatchBody": {
        "additionalProperties": false,
        "properties": {
          "grid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GridConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "widgets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WidgetInstance"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widgets"
          }
        },
        "title": "LayoutPatchBody",
        "type": "object"
      },
      "LayoutPutBody": {
        "additionalProperties": false,
        "properties": {
          "grid": {
            "$ref": "#/components/schemas/GridConfig"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/WidgetInstance"
            },
            "title": "Widgets",
            "type": "array"
          }
        },
        "required": [
          "widgets"
        ],
        "title": "LayoutPutBody",
        "type": "object"
      },
      "LayoutResponse": {
        "additionalProperties": false,
        "properties": {
          "grid": {
            "$ref": "#/components/schemas/GridConfig"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "pageId": {
            "title": "Pageid",
            "type": "string"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/WidgetInstance"
            },
            "title": "Widgets",
            "type": "array"
          }
        },
        "required": [
          "pageId"
        ],
        "title": "LayoutResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WidgetInstance": {
        "additionalProperties": false,
        "properties": {
          "h": {
            "title": "H",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "props": {
            "additionalProperties": true,
            "title": "Props",
            "type": "object"
          },
          "w": {
            "title": "W",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "WidgetInstance",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Forza Racer Telemetry Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/layouts/{page_id}": {
      "get": {
        "operationId": "get_layout_api_layouts__page_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Layout",
        "tags": [
          "layouts"
        ]
      },
      "patch": {
        "operationId": "patch_layout_api_layouts__page_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutPatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Layout",
        "tags": [
          "layouts"
        ]
      },
      "put": {
        "operationId": "put_layout_api_layouts__page_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutPutBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Layout",
        "tags": [
          "layouts"
        ]
      }
    },
...
  }
}